Exercise 4.20
Exercise 4.20: Bloglist expansion step 8. Refactoring the application in such a way that the token is taken to a middleware. Exercise 4.21 is also refactored so as to access the token with request.token

+++ blogposts/controllers/blogs.js
@@ -1,27 +1,20 @@
+const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
-const jwt = require('jsonwebtoken')
-
-const getTokenFrom = request => {
-    const authorization = request.get('authorization')
-    if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-        return authorization.substring(7)
-    }
-    return null
-}
 
 blogsRouter.get('/', async (request, response) => {
+    
     const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs.map(blog => blog.toJSON()))
 })
 
 blogsRouter.post('/', async (request, response) => {
+
     const body = request.body
-    const token = getTokenFrom(request)
-    const decodedToken = jwt.verify(token, process.env.SECRET)
-    //    console.log(decodedToken)
-    if (!token || !decodedToken.id) {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!request.token || !decodedToken.id) {
         return response.status(401).json({ error: 'token missing or invalid' })
     }
     const user = await User.findById(decodedToken.id)
@@ -42,26 +35,27 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    const token = getTokenFrom(request)
-    const decodedToken = jwt.verify(token, process.env.SECRET)
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
     const userid = decodedToken.id
-    if (!token || !decodedToken.id) {
+    if (!request.token || !decodedToken.id) {
         return response.status(401).json({ error: 'token missing or invalid' })
     }
     const blog = await Blog.findById(request.params.id)
-    if (blog.user.toString() === userid.toString()) {
+    if (blog && blog.user.toString() === userid.toString()) {
         const blogToDelete = await Blog.findByIdAndRemove(request.params.id)
         if (blogToDelete) {
             response.status(204).end()
-        } else {
-            response.status(404).send(`The data is already deleted from database`)
         }
+    } else if (!blog) {
+        response.status(404).send('The data is already deleted from server')
     } else {
-        response.status(404).send('Permission to delete the data is denied')
+        response.status(401).end()
     }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
+
     const body = request.body
 
     const blog = {
